backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
psycopg[binary]==3.1.13
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2